+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -1,3 +1,8 @@
+## NEXT
+
+* Updates minimum supported SDK version to Flutter 3.13/Dart 3.1.
+* Updates compileSdk to 34.
+
 ## 3.15.0
 
 * Adds support for `setOnScrollPositionChange` method to the `AndroidWebViewController`.